# docs/swagger.yaml
basePath: /
definitions:
  handlers.LoginRequest:
    properties:
      idToken:
        type: string
    required:
    - idToken
    type: object
...
  types.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
host: provolo-backend.onrender.com
info:
  contact: {}
  description: This is the Provolo backend API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Provolo API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user with Firebase ID token and creates session cookie
      parameters:
      - description: Login request with ID token
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: User login
      tags:
      - auth
...
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Checks if user exists in Firestore and creates profile with starter
        tier if not
      parameters:
      - description: Firebase ID token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: User signup or ensure existence
      tags:
      - auth
  /api/v1/auth/verify:
    get:
      description: Verifies the current user session and returns user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Verify user session
      tags:
      - auth
...
  /api/v1/protected/profile:
    get:
      description: Retrieves the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.APIResponse'
      summary: Get user profile
      tags:
      - auth
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"